Code refactor and dynamic navbar changes
A lot of restructuring. Different routes now have their own file in the routes folder. One for campgrounds, comments and index (which contains the auth routes). Changed up a few require statements. More were needed in app.js. Navbar now displays different links depending on whether a user is signed in. Also displays the name of the signed in user