Added authorization for comments and refactored middleware
Update and delete comments can only be done if the author of said comments is signed in. Buttons are hidden and routes are blocked using middleware when the author is not signed in. Middleware for comment and campground routes which authorize routes and functions are moved to middleware directory, index.js file and exported and required where neccessary.