Added flash messages and fixed a bug
Used connect-flash npm along with bootstrap alert components to give meaningful feedback to users when an error occurs. The bug fix is added error checking for when a user might manually insert a valid but non-existent mongodb id and tries to post it. An extra condition in the if statements of the middleware and all the routes should fix that