Added price to campgrounds and revamped landing page
Price field now added to the campground schema and relevant form inputs have been added. Price is also displayed on the show page. If null or zero, appropriate message is displayed. Landing page looks like an actual landing page. A slideshow with cross-fading pictures of camps is now showing with a button that takes the user to all campgrounds. Footer is also updated with links to view all and to my portfolio page