Added form validation and collapsable navbar
Fixed a problem where a user to enter a blank camp or comment by adding the 'required' attribute to all form inputs. Used bootstrap classes to make the mavbar collapsable and have a burger icon for smaller screens. Also moved some styles from the camp index page to the CSS file and updated the flex items to align properly on smaller devices.